+++ weblate/formats/models.py
@@ -147,6 +147,7 @@ class FormatsConf(AppConf):
         "weblate.formats.ttkit.XWikiFullPageFormat",
         "weblate.formats.ttkit.TBXFormat",
         "weblate.formats.ttkit.StringsdictFormat",
+        "weblate.formats.ttkit.FluentFormat",
     )
 
     class Meta: